More flexible resource pools
- Allow user defined pools.
- Provide combinators for pools (`pair`, `both`, `either`).
- Implement waiting for confirmation using the new API.

Before, if you set the threshold to allow a job to start, and then set
it back again before the job actually started (because it was waiting
for some other resource), the job would still start once the resource
was available. Now jobs can only start when their resource is available
and the confirmation threshold allows it at the same time.

This means that setting the threshold to confirm everything will bring
things to a stop much more quickly, since it affects all queued jobs.